Remove GeneratorSettingsFix as it's not needed.
I misunderstood the code I wrote for it and thoughjt the walker for it actually does
anything useful, but the only thing it did was check for world type, and give the whole
level nbt to the data fixer anyway. To avoid creating a different fix type for each
world type, code using that would still have to check for it's world type anyway, which
makes this class completely unnecessary.